+++ libensemble/__init__.py
@@ -10,7 +10,7 @@ __author__ = "Jeffrey Larson, Stephen Hudson, Stefan M. Wild, David Bindel and J
 __credits__ = "Argonne National Laboratory"
 
 from libensemble import logger
-from libensemble.utils.pydantic_support import pydanticV1, pydanticV2
+from libensemble.utils.pydantic_bindings import pydanticV1, pydanticV2
 
 if pydanticV1:
     from .specs import specsV1 as specs
